made the recommended changes
